created navbar file, and moved some containers files into component folder
