+++ src/containers/HomeConnected.js
@@ -1,11 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import lexusLogo from '../img/Lexus.svg';
 import Cars from '../components/Cars';
 import Car from '../components/Car';
 import AppointmentForm from '../components/AppointmentForm';
 import Appointments from './Appointments';
-import './HomeConnected.scss';
+import SideMenu from '../components/SideMenu';
 
 class HomeConnected extends Component {
   constructor(props) {
@@ -13,6 +12,7 @@ class HomeConnected extends Component {
     this.state = {
       selection: 'SEDAN',
     };
+    this.select = this.select.bind(this);
     this.clickHandler = this.clickHandler.bind(this);
     this.makeAppointment = this.makeAppointment.bind(this);
     this.updateSelection = this.updateSelection.bind(this);
@@ -21,8 +21,8 @@ class HomeConnected extends Component {
   }
 
   // eslint-disable-next-line class-methods-use-this
-  removeClassTo() {
-    const menuList = document.getElementsByClassName('menu-list')[0];
+  removeClassTo(menu) {
+    const menuList = menu;
     for (let i = 0; i < menuList.children.length; i += 1) {
       menuList.children[i].classList.remove('selection');
     }
@@ -77,7 +77,7 @@ class HomeConnected extends Component {
   }
 
   select(e) {
-    this.removeClassTo();
+    this.removeClassTo(e.target.parentElement);
     const item = e.target;
     this.addClassTo(item);
     if (item.innerHTML === 'SEDAN') {
@@ -108,34 +108,10 @@ class HomeConnected extends Component {
   }
 
   render() {
-    const { handleClick } = this.props;
+    const { logOutAction } = this.props;
     return (
       <div className="container-4">
-        <div className="side-menu">
-          <div className="logo">
-            <img src={lexusLogo} alt="Lexus" />
-          </div>
-          <div className="menu">
-            <nav>
-              <ul className="menu-list">
-                <li onClick={this.select.bind(this)} role="presentation" className="selection">SEDAN</li>
-                <li onClick={this.select.bind(this)} role="presentation">SUV</li>
-                <li onClick={this.select.bind(this)} role="presentation">COUPE</li>
-                <li onClick={this.select.bind(this)} role="presentation">HYBRID</li>
-                <li onClick={this.select.bind(this)} role="presentation" id="appointments">APPOINTMENTS</li>
-              </ul>
-            </nav>
-          </div>
-          <div className="social-media">
-            <div className="icons">
-              <i className="fab fa-twitter" />
-              <i className="fab fa-facebook-f" />
-              <i className="fab fa-google-plus-g" />
-              <i className="fab fa-pinterest-p" />
-            </div>
-            <button type="button" onClick={handleClick}>Log Out</button>
-          </div>
-        </div>
+        <SideMenu handleSelection={this.select} handleClick={logOutAction} />
         <div className="">{this.content()}</div>
       </div>
     );
@@ -143,7 +119,7 @@ class HomeConnected extends Component {
 }
 
 HomeConnected.propTypes = {
-  handleClick: PropTypes.func.isRequired,
+  logOutAction: PropTypes.func.isRequired,
 };
 
 export default HomeConnected;
