+++ src/index.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import App from './components/App';
 import store from './reducers/index';
 import * as serviceWorker from './serviceWorker';
 import './index.scss';
+import App from './containers/App';
 
 ReactDOM.render(
   <React.StrictMode>
